Adding Authentication

+++ survey-buzz/src/app/components/survey-list/survey-list.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CrudService } from './../../service/crud.service';
+import { AuthService } from './../../shared/auth.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-survey-list',
@@ -8,8 +10,13 @@ import { CrudService } from './../../service/crud.service';
 })
 export class SurveyListComponent implements OnInit {
   Surveys: any = [];
+  currentUser: any = {};
 
-  constructor(private crudService: CrudService) {}
+  constructor(
+    private crudService: CrudService,
+    public authService: AuthService,
+    public router: Router
+    ) {}
 
   ngOnInit(): void {
     this.crudService.GetSurveys().subscribe((res) => {
@@ -20,10 +27,14 @@ export class SurveyListComponent implements OnInit {
 
   delete(id: any, i: any) {
     console.log(id);
-    if (window.confirm('Do you want to go ahead?')) {
-      this.crudService.deleteSurvey(id).subscribe((res) => {
-        this.Surveys.splice(i, 1);
-      });
-    }
+    if (this.authService.isLoggedIn) {
+      if (window.confirm('Do you want to go ahead?')) {
+        this.crudService.deleteSurvey(id).subscribe((res) => {
+          this.Surveys.splice(i, 1);
+        });
+      }
+    } else {
+      this.router.navigate(['log-in'])
+    } 
   }
 }
